chore: amend README
