+++ aidial_adapter_openai/audio_api/adapter.py
@@ -1,10 +1,10 @@
 import base64
-from typing import Any, AsyncIterator, Literal
+from typing import Any, AsyncIterator
 
 from aidial_sdk.exceptions import RequestValidationError
 from openai import AsyncAzureOpenAI, AsyncOpenAI
-from pydantic import BaseModel, Field
 
+from aidial_adapter_openai.audio_api.configuration import Configuration
 from aidial_adapter_openai.dial_api.attachment import (
     upload_message_attachments_to_storage,
 )
@@ -48,50 +48,6 @@ def collect_system_messages(messages: list[dict]) -> str | None:
         if message.get("role") in ("system", "developer"):
             ret += collect_message_text_content(message)
     return ret.strip() or None
-
-
-Voices = Literal[
-    "alloy",
-    "ash",
-    "ballad",
-    "coral",
-    "echo",
-    "fable",
-    "onyx",
-    "nova",
-    "sage",
-    "shimmer",
-    "verse",
-]
-
-Formats = Literal["mp3", "opus", "aac", "flac", "wav", "pcm"]
-
-
-class Configuration(BaseModel):
-    instructions: str | None = Field(
-        default=None,
-        description=(
-            "Control the voice of your generated audio with additional instructions. "
-            "Does not work with `tts-1` or `tts-1-hd`. "
-            "The instruction from the system and developer messages "
-            "will be attached to the instructions from the configuration."
-        ),
-    )
-    voice: str | Voices | None = Field(
-        default="alloy",
-        description="The voice to use when generating the audio.",
-    )
-    speed: float | None = Field(
-        default=None,
-        description=(
-            "The speed of the generated audio. "
-            "Select a value from `0.25` to `4.0`. `1.0` is the default. "
-            "Does not work with `gpt-4o-mini-tts`."
-        ),
-    )
-    response_format: str | Formats | None = Field(
-        default=None, description="The format of the generated audio."
-    )
 
 
 async def chat_completion(
