+++ aidial_adapter_openai/configuration/deployment_type.py
@@ -12,4 +12,4 @@ class ChatCompletionDeploymentType(StrEnum):
     RESPONSES_API = "RESPONSES_API"
     COMPLETIONS_API = "COMPLETIONS_API"
     AZURE_VIDEO_API = "VIDEO_API"
-    SPEECH_API = "SPEECH_API"
+    AUDIO_SPEECH_API = "SPEECH_API"
